One Program == Many Sketches
LCD menu with a potentiometer used to to select up to 10 Sketches.  menuOne function has a back and forth led scanner as an example on how to use this program to copy and paste sketches.  I find it is easier to initialize the variable into the function that uses it.  This keeps variables local and easier to keep up with.